Refactor Dockerfile and implement Flask app for video downloading functionality with MP3 and MP4 support; add index.html for user interface and update requirements.txt
