+++ requirements.txt
@@ -1,3 +1,3 @@
-Flask==2.0.3
-gunicorn==20.1.0
-yt-dlp==2021.12.1
+gunicorn
+flask
+yt-dlp
